moved functions to script_functions
